+++ src/olmo_core/nn/transformer/model.py
@@ -1,4 +1,5 @@
 import logging
+from collections import defaultdict
 from functools import cached_property
 from typing import (
     TYPE_CHECKING,
@@ -197,14 +198,17 @@ class Transformer(nn.Module):
 
     def get_rope_buffers(
         self, seq_len: int, device: Optional[torch.device] = None
-    ) -> Optional[RoPEBuffers]:
+    ) -> Dict[int, Optional[RoPEBuffers]]:
+        """
+        Get the RoPE buffers to pass to each layer.
+        """
         if device is None:
             device = self.device
-        for block in self.blocks.values():
+        rope_buffers = {}
+        for key, block in self.blocks.items():
             rope = cast(Optional[RotaryEmbeddingBase], block.attention.rope)  # type: ignore
-            if rope is not None:
-                return rope.get_buffers(seq_len, device)
-        return None
+            rope_buffers[int(key)] = None if rope is None else rope.get_buffers(seq_len, device)
+        return rope_buffers
 
     @torch.no_grad()
     def init_weights(
@@ -305,13 +309,19 @@ class Transformer(nn.Module):
         loss_div_factor: Optional[Union[torch.Tensor, float]] = None,
         return_logits: Optional[bool] = None,
         **kwargs,
-    ) -> Tuple[torch.Tensor, Optional[torch.Tensor], Dict[str, Any], Dict[str, Any]]:
+    ) -> Tuple[
+        torch.Tensor,
+        Optional[torch.Tensor],
+        Dict[str, Any],
+        Dict[int, Dict[str, Any]],
+        Dict[str, Any],
+    ]:
         # NOTE: with pipeline parallelism input_ids might actually be an intermediate output,
         # so we have to be careful here.
         B, S = input_ids.shape[:2]
 
-        block_kwargs: Dict[str, Any] = {}
-
+        all_block_kwargs: Dict[str, Any] = {}
+        per_block_kwargs: Dict[int, Dict[str, Any]] = defaultdict(dict)
         lm_head_kwargs: Dict[str, Any] = dict(
             ignore_index=ignore_index,
             loss_reduction=loss_reduction,
@@ -322,7 +332,7 @@ class Transformer(nn.Module):
         if loss_div_factor is not None:
             loss_div_factor = move_to_device(loss_div_factor, self.device)
             lm_head_kwargs["loss_div_factor"] = loss_div_factor
-            block_kwargs["loss_div_factor"] = loss_div_factor
+            all_block_kwargs["loss_div_factor"] = loss_div_factor
 
         # Prepare document length inputs.
         max_doc_len: Optional[int] = None
@@ -341,23 +351,23 @@ class Transformer(nn.Module):
             keys = ["input_ids"]
 
             # NOTE: initialize buffer(s) on CPU to avoid possible host-device sync when sharding.
-            rope_buffers = self.get_rope_buffers(S, torch.device("cpu"))
-            if rope_buffers is not None:
-                if rope_buffers.pos_sin is not None:
-                    inputs.append(rope_buffers.pos_sin)
-                    seq_dims.append(0)
-                    pad_values.append(0.0)
-                    keys.append("pos_sin")
-                if rope_buffers.pos_cos is not None:
-                    inputs.append(rope_buffers.pos_cos)
-                    seq_dims.append(0)
-                    pad_values.append(0.0)
-                    keys.append("pos_cos")
-                if rope_buffers.freqs_cis is not None:
-                    inputs.append(rope_buffers.freqs_cis)
-                    seq_dims.append(0)
-                    pad_values.append(0.0)
-                    keys.append("freqs_cis")
+            for block_idx, rope_buffers in self.get_rope_buffers(S, torch.device("cpu")).items():
+                if rope_buffers is not None:
+                    if rope_buffers.pos_sin is not None:
+                        inputs.append(rope_buffers.pos_sin)
+                        seq_dims.append(0)
+                        pad_values.append(0.0)
+                        keys.append(f"block_{block_idx}.pos_sin")
+                    if rope_buffers.pos_cos is not None:
+                        inputs.append(rope_buffers.pos_cos)
+                        seq_dims.append(0)
+                        pad_values.append(0.0)
+                        keys.append(f"block_{block_idx}.pos_cos")
+                    if rope_buffers.freqs_cis is not None:
+                        inputs.append(rope_buffers.freqs_cis)
+                        seq_dims.append(0)
+                        pad_values.append(0.0)
+                        keys.append(f"block_{block_idx}.freqs_cis")
 
             if labels is not None:
                 inputs.append(labels)
@@ -384,7 +394,8 @@ class Transformer(nn.Module):
                     length_multiple=16,
                 )
                 for key, value in additional_inputs.items():
-                    block_kwargs[key] = move_to_device(value, self.device)
+                    all_block_kwargs[key] = move_to_device(value, self.device)
+
             else:
                 inputs = cp_load_balancer.batch_shard(
                     inputs=inputs,
@@ -393,20 +404,26 @@ class Transformer(nn.Module):
                 )
 
             for key, value in zip(keys, inputs):
-                block_kwargs[key] = move_to_device(value, self.device)
+                if key.startswith("block_"):
+                    block_key, key = key.split(".", 1)
+                    block_idx = int(block_key.replace("block_", ""))
+                    per_block_kwargs[block_idx][key] = move_to_device(value, self.device)
+                else:
+                    all_block_kwargs[key] = move_to_device(value, self.device)
 
-            input_ids = block_kwargs.pop("input_ids")
-            labels = block_kwargs.pop("labels", None)
+            input_ids = all_block_kwargs.pop("input_ids")
+            labels = all_block_kwargs.pop("labels", None)
         else:
             input_ids = move_to_device(input_ids, self.device)
             labels = move_to_device(labels, self.device)
-            block_kwargs["max_doc_len"] = max_doc_len
-            block_kwargs["cu_doc_lens"] = move_to_device(cu_doc_lens, self.device)
+            all_block_kwargs["max_doc_len"] = max_doc_len
+            all_block_kwargs["cu_doc_lens"] = move_to_device(cu_doc_lens, self.device)
 
         return (
             input_ids,
             labels,
-            block_kwargs,
+            all_block_kwargs,
+            per_block_kwargs,
             lm_head_kwargs,
         )
 
@@ -429,7 +446,13 @@ class Transformer(nn.Module):
 
         :returns: The logits if ``labels`` is ``None`` or the losses if ``labels`` is not ``None``.
         """
-        input_ids, labels, block_kwargs, lm_head_kwargs = self._prepare_inputs(
+        (
+            input_ids,
+            labels,
+            all_block_kwargs,
+            per_block_kwargs,
+            lm_head_kwargs,
+        ) = self._prepare_inputs(
             input_ids,
             labels,
             ignore_index=ignore_index,
@@ -445,11 +468,13 @@ class Transformer(nn.Module):
         h = self.embeddings(input_ids) if self.embeddings is not None else input_ids
 
         # Run each block.
-        for block in self.blocks.values():
+        for block_key, block in self.blocks.items():
+            block_idx = int(block_key)
+            block_kwargs = per_block_kwargs.get(block_idx, {})
             # Mark sizes as dynamic for torch.compile().
             if self.compile_enabled:
                 mark_dynamic(h, (0, 1), strict=False)
-            h = block(h, **block_kwargs)
+            h = block(h, **all_block_kwargs, **block_kwargs)
 
         # Get final logits but again pass-through in case of pipeline parallelism.
         if self.lm_head is not None: